Ignore non-JS files in destinationDir
Some operating systems silently and automatically create "hidden files" inside new directories. So, for example on Mac OS, the destinationDir can have a `.DS_Store` file appear within it from the system. The code in point-of-view that tries to `require` compiled templates should ignore these, and indeed any files that do not end in `.js`.