Create a precise cache key for partials for LRU
The key has to consider `page`, `partials` and `requestedPath` to be
usable. Ignoring any of these values will result in a key that is too
broad, which will result in invalid cache hits.

Note: because `requestedPath` is used for determining whether
minification should be used on the content, it also has to be used in
the cache key.

As below

```js
if (useHtmlMinification(globalOptions, requestedPath)) {
  data = globalOptions.useHtmlMinifier.minify(data, globalOptions.htmlMinifierOptions || {})
}
```

**Before**

| variable | value |
|----------|-------|
| page     | /templates/index.mustache |
| partials | { content: '/templates/some-partial.mustache' } |
| requestedPath | /some-path |

Would result in a cache key of `/templates/index.mustache-Partials`,
which completely ignores the partials and the requested path

If the next request would use the same template with different partials,
it would get the same cache key, and thus the partials from the first request.

**After**

| variable | value |
|----------|-------|
| page     | /templates/index.mustache |
| partials | { content: '/templates/some-partial.mustache' } |
| requestedPath | /some-path |

Would result in a cache key of
`/templates/index.mustache|content:/templates/some-partial.mustache|/some-path-Partials`,
which clearly uses `page`, `partials` and `requestedPath` for generating
the key.