fix: stop duplication of path in template dir (#440)
Closes #437

When the `root` option is passed explicitly and the `art-template` engine is in use, the `root` value is being duplicated in the resolved template directory.

This leads to an error like this: notice how `templates` is present twice in the path.

```
  -{"statusCode":500,"error":"Internal Server Error","message":"template not found: ENOENT: no such file or directory, open
  '/home/rich/point-of-view/templates/templates/index.art'"}
```

The fix is to remove this duplication. The `resolveTemplateDir` function already uses any `root` option (if present) so there's no need to consider it again.